Separate AbstractDataSourcePreparer
